Extract shared vocoder and routed tensor computation
